chore: add prior youtube tutorial repos
The existing repos for prior YouTube tutorials are
being committed to this newly created Git repo to
house all repos created for different YouTube
tutorials going forward.

This commit has been marked as chore as this repo
would act as a place to organize other Git repos
with code pertaining to tutorials. So, there will
not be any code directly associated to this repo
and will not have commits marked as: feat, fix,
refactor, with reference to the Conventional Commits
Specification.